added balance sheet and cashflow api

+++ src/main/java/cn/initialfire/tushare/api/stock/responsedata/StockIncomeReport.java
@@ -45,437 +45,437 @@ public class StockIncomeReport {
     /**
      * 基本每股收益
      */
-    private Double basicEps;
+    private Float basicEps;
 
     /**
      * 稀释每股收益
      */
-    private Double dilutedEps;
+    private Float dilutedEps;
 
     /**
      * 营业总收入
      */
-    private Double totalRevenue;
+    private Float totalRevenue;
 
     /**
      * 营业收入
      */
-    private Double revenue;
+    private Float revenue;
 
     /**
      * 营业总成本
      */
-    private Double totalCogs;
+    private Float totalCogs;
 
     /**
      * 营业成本
      */
-    private Double operCost;
+    private Float operCost;
 
     /**
      * 营业总成本（二）
      */
-    private Double totalOpcost;
+    private Float totalOpcost;
 
     /**
      * 营业支出
      */
-    private Double operExp;
+    private Float operExp;
 
     /**
      * 销售费用
      */
-    private Double sellExp;
+    private Float sellExp;
 
     /**
      * 研发费用
      */
-    private Double rdExp;
+    private Float rdExp;
 
     /**
      * 管理费用
      */
-    private Double adminExp;
+    private Float adminExp;
 
     /**
      * 财务费用
      */
-    private Double finExp;
+    private Float finExp;
 
     /**
      * 财务费用:利息费用
      */
-    private Double finExpIntExp;
+    private Float finExpIntExp;
 
     /**
      * 财务费用:利息收入
      */
-    private Double finExpIntInc;
+    private Float finExpIntInc;
 
     /**
      * 利息收入
      */
-    private Double intIncome;
+    private Float intIncome;
 
     /**
      * 利息支出
      */
-    private Double intExp;
+    private Float intExp;
 
     /**
      * 资产减值损失
      */
-    private Double assetsImpairLoss;
+    private Float assetsImpairLoss;
 
     /**
      * 其他资产减值损失
      */
-    private Double othImpairLossAssets;
+    private Float othImpairLossAssets;
 
     /**
      * 其他经营净收益
      */
-    private Double nOthIncome;
+    private Float nOthIncome;
 
     /**
      * 其他业务净收益
      */
-    private Double nOthBIncome;
+    private Float nOthBIncome;
 
     /**
      * 其他业务收入
      */
-    private Double othBIncome;
+    private Float othBIncome;
 
     /**
      * 其他业务成本
      */
-    private Double otherBusCost;
+    private Float otherBusCost;
 
     /**
      * 投资收益
      */
-    private Double investIncome;
+    private Float investIncome;
 
     /**
      * 资产处置收益
      */
-    private Double assImpGain;
+    private Float assImpGain;
 
     /**
      * 非流动资产处置损失
      */
-    private Double ncaDisploss;
+    private Float ncaDisploss;
 
     /**
      * 汇兑收益
      */
-    private Double forexGain;
+    private Float forexGain;
 
     /**
      * 息税前利润
      */
-    private Double ebit;
+    private Float ebit;
 
     /**
      * 息税折旧摊销前利润
      */
-    private Double ebitda;
+    private Float ebitda;
 
     /**
      * 所得税费用
      */
-    private Double incomeTax;
+    private Float incomeTax;
 
     /**
      * 营业税金及附加
      */
-    private Double bizTaxSurchg;
+    private Float bizTaxSurchg;
 
     /**
      * 营业利润
      */
-    private Double operateProfit;
+    private Float operateProfit;
 
     /**
      * 营业外收入
      */
-    private Double nonOperIncome;
+    private Float nonOperIncome;
 
     /**
      * 营业外支出
      */
-    private Double nonOperExp;
+    private Float nonOperExp;
 
     /**
      * 利润总额
      */
-    private Double totalProfit;
+    private Float totalProfit;
 
     /**
      * 净利润(含少数股东损益)
      */
-    private Double nIncome;
+    private Float nIncome;
 
     /**
      * 净利润(不含少数股东损益)
      */
-    private Double nIncomeAttrP;
+    private Float nIncomeAttrP;
 
     /**
      * 扣除非经常性损益后的净利润（更正前）
      */
-    private Double netAfterNrLpCorrect;
+    private Float netAfterNrLpCorrect;
 
     /**
      * 少数股东损益
      */
-    private Double minorityGain;
+    private Float minorityGain;
 
     /**
      * 其他综合收益
      */
-    private Double othComprIncome;
+    private Float othComprIncome;
 
     /**
      * 综合收益总额
      */
-    private Double tComprIncome;
+    private Float tComprIncome;
 
     /**
      * 归属于母公司(或股东)的综合收益总额
      */
-    private Double comprIncAttrP;
+    private Float comprIncAttrP;
 
     /**
      * 归属于少数股东的综合收益总额
      */
-    private Double comprIncAttrMs;
+    private Float comprIncAttrMs;
 
     /**
      * 可供分配利润
      */
-    private Double distableProfit;
+    private Float distableProfit;
 
     /**
      * 保险业务支出
      */
-    private Double insuranceExp;
+    private Float insuranceExp;
 
     /**
      * 未分配利润
      */
-    private Double undistProfit;
+    private Float undistProfit;
 
     /**
      * 信用减值损失
      */
-    private Double creditImpaLoss;
+    private Float creditImpaLoss;
 
     /**
      * 其他收益
      */
-    private Double othIncome;
+    private Float othIncome;
 
     /**
      * 资产处置收益
      */
-    private Double assetDispIncome;
+    private Float assetDispIncome;
 
     /**
      * 持续经营净利润
      */
-    private Double continuedNetProfit;
+    private Float continuedNetProfit;
 
     /**
      * 终止经营净利润
      */
-    private Double endNetProfit;
+    private Float endNetProfit;
 
     /**
      * 对联营企业和合营企业的投资收益
      */
-    private Double assInvestIncome;
+    private Float assInvestIncome;
 
     /**
      * 已赚保费
      */
-    private Double premEarned;
+    private Float premEarned;
 
     /**
      * 手续费及佣金收入
      */
-    private Double commIncome;
+    private Float commIncome;
 
     /**
      * 手续费及佣金净收入
      */
-    private Double nCommisIncome;
+    private Float nCommisIncome;
 
     /**
      * 保险业务收入
      */
-    private Double premIncome;
+    private Float premIncome;
 
     /**
      * 分出保费
      */
-    private Double outPrem;
+    private Float outPrem;
 
     /**
      * 提取未到期责任准备金
      */
-    private Double unePremReser;
+    private Float unePremReser;
 
     /**
      * 分保费收入
      */
-    private Double reinsIncome;
+    private Float reinsIncome;
 
     /**
      * 证券承销业务净收入
      */
-    private Double nSecTbIncome;
+    private Float nSecTbIncome;
 
     /**
      * 证券承销净收益
      */
-    private Double nSecUwIncome;
+    private Float nSecUwIncome;
 
     /**
      * 受托客户资产管理业务净收入
      */
-    private Double nAssetMgIncome;
+    private Float nAssetMgIncome;
 
     /**
      * 以公允价值计量且其变动计入当期损益的金融资产净收益
      */
-    private Double fvValueChgGain;
+    private Float fvValueChgGain;
 
     /**
      * 手续费及佣金支出
      */
-    private Double commExp;
+    private Float commExp;
 
     /**
      * 退保金
      */
-    private Double premRefund;
+    private Float premRefund;
 
     /**
      * 赔付总支出
      */
-    private Double compensPayout;
+    private Float compensPayout;
 
     /**
      * 提取保险责任准备金
      */
-    private Double reserInsurLiab;
+    private Float reserInsurLiab;
 
     /**
      * 保户红利支出
      */
-    private Double divPayt;
+    private Float divPayt;
 
     /**
      * 分保费用
      */
-    private Double reinsExp;
+    private Float reinsExp;
 
     /**
      * 摊回赔付支出
      */
-    private Double compensPayoutRefu;
+    private Float compensPayoutRefu;
 
     /**
      * 摊回保险责任准备金
      */
-    private Double insurReserRefu;
+    private Float insurReserRefu;
 
     /**
      * 摊回分保费用
      */
-    private Double reinsCostRefund;
+    private Float reinsCostRefund;
 
     /**
      * 盈余公积转入
      */
-    private Double transferSurplusRese;
+    private Float transferSurplusRese;
 
     /**
      * 住房周转金转入
      */
-    private Double transferHousingImpre;
+    private Float transferHousingImpre;
 
     /**
      * 其他转入
      */
-    private Double transferOth;
+    private Float transferOth;
 
     /**
      * 调整以前年度损益
      */
-    private Double adjLossgain;
+    private Float adjLossgain;
 
     /**
      * 提取法定盈余公积
      */
-    private Double withdraLegalSurplus;
+    private Float withdraLegalSurplus;
 
     /**
      * 提取法定公益金
      */
-    private Double withdraLegalPubfund;
+    private Float withdraLegalPubfund;
 
     /**
      * 提取企业发展基金
      */
-    private Double withdraBizDevfund;
+    private Float withdraBizDevfund;
 
     /**
      * 提取储备基金
      */
-    private Double withdraReseFund;
+    private Float withdraReseFund;
 
     /**
      * 提取任意盈余公积金
      */
-    private Double withdraOthErsu;
+    private Float withdraOthErsu;
 
     /**
      * 职工奖金福利
      */
-    private Double workersWelfare;
+    private Float workersWelfare;
 
     /**
      * 可分配给股东的利润
      */
-    private Double distrProfitShrhder;
+    private Float distrProfitShrhder;
 
     /**
      * 应付优先股股利
      */
-    private Double prfsharePayableDvd;
+    private Float prfsharePayableDvd;
 
     /**
      * 应付普通股股利
      */
-    private Double comsharePayableDvd;
+    private Float comsharePayableDvd;
 
     /**
      * 转作股本的普通股股利
      */
-    private Double capitComstockDiv;
+    private Float capitComstockDiv;
 
     /**
      * 套期保值业务净收益
      */
-    private Double netExpoHedgingBenefits;
+    private Float netExpoHedgingBenefits;
 
     /**
      * 以摊余成本计量的金融资产终止确认收益
      */
-    private Double amodcostFinAssets;
+    private Float amodcostFinAssets;
 
     /**
      * 更新标识
